Refactor DropReadwriteSplittingRuleStatementUpdaterTest
